requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
websockets>=12.0
aiosqlite>=0.19.0
pyserial-asyncio>=0.6
tomli>=2.0.0;python_version<"3.11"
aprslib>=0.7.2
geopy>=2.4.0
